+++ pycln/cli.py
@@ -14,6 +14,21 @@ app = typer.Typer(name=__name__, add_completion=True)
 @app.command(context_settings=dict(help_option_names=["-h", "--help"]))
 def main(  # pylint: disable=R0913,R0914
     paths: List[Path] = typer.Argument(None, help="Directories or files paths."),
+    skip_imports: List[str] = typer.Option(
+        [],
+        "--skip-imports",
+        show_default=False,
+        help=(
+            "A list of module/package names to be skipped globally."
+            " This works similar to `# nopycln: import`"
+            " (to skip certain import statements)"
+            " but at a global scope where you provide a comma separated list"
+            " of module/package/library names"
+            " to be skipped via the CLI option like `--skip-imports os,time,pycln`"
+            " or by passing a list via a config file"
+            " such as `skip_imports: [os, time, pycln]` (recommended)."
+        ),
+    ),
     config: Optional[Path] = typer.Option(
         None,
         "--config",
@@ -141,6 +156,7 @@ def main(  # pylint: disable=R0913,R0914
 ):
     configs = Config(
         paths=paths,
+        skip_imports=set(skip_imports),
         config=config,
         include=include,  # type: ignore
         exclude=exclude,  # type: ignore